backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
langchain-community
langchain-huggingface
langchain
langdetect
google-generativeai
sentence-transformers
faiss-cpu
fastapi[standard]